fix: float to integer

See merge request s08-ai-speech-sub2/S08P22B202!314

+++ machine/TTS/app-test.py
@@ -63,7 +63,7 @@ def tts(text: str, cons_num: int = -1):
 
     # melgan-eval
     samplerate = 22050
-    audioinit = samplerate*0.4
+    audioinit = int(samplerate*0.4)
     audiobasic = samplerate*3
     g_audio = vocoder(mel)
     g_audio = g_audio.squeeze().cpu()
